Extract `expression` and `rule` packages
