+++ java/src/main/kotlin/io/spine/validation/java/option/setonce/SetOnceBooleanField.kt
@@ -30,11 +30,8 @@ import com.intellij.psi.PsiClass
 import io.spine.protodata.ast.Field
 import io.spine.protodata.ast.PrimitiveType
 import io.spine.protodata.java.Expression
-import io.spine.protodata.java.ClassName
-import io.spine.protodata.java.call
 import io.spine.protodata.type.TypeSystem
 import io.spine.tools.psi.java.method
-import io.spine.validation.java.StringClass
 
 /**
  * Renders Java code to support `(set_once)` option for the given boolean [field].
