+++ java/src/main/kotlin/io/spine/validation/java/rule/CustomRuleGenerators.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.rule
 
 import com.google.protobuf.Message
 import io.spine.protobuf.unpackGuessingType
@@ -33,6 +33,11 @@ import io.spine.validation.RecursiveValidation
 import io.spine.validation.RequiredOneof
 import io.spine.validation.isMessageWide
 import io.spine.validation.isSimple
+import io.spine.validation.java.GenerationContext
+import io.spine.validation.java.RequiredOneofGenerator
+import io.spine.validation.java.UnsupportedRuleGenerator
+import io.spine.validation.java.ValidateGenerator
+import io.spine.validation.java.inTimeGenerator
 
 /**
  * Creates a [CodeGenerator] for a custom validation operator for the given context.
